Cast MS_YEAR as it needs to be

+++ ms.c
@@ -26,7 +26,7 @@
 #define MS_HOUR 60 * MS_MIN
 #define MS_DAY 24 * MS_HOUR
 #define MS_WEEK 7 * MS_DAY
-#define MS_YEAR 52 * MS_WEEK
+#define MS_YEAR 52 * (long long)MS_WEEK
 
 /*
  * Convert the given `str` representation to microseconds,
